+++ perf/tests/test_tools.py
@@ -1,6 +1,4 @@
 import datetime
-import os.path
-import sys
 
 import six
 
@@ -9,7 +7,6 @@ from perf._formatter import (format_filesize, format_seconds, format_timedelta,
                              format_timedeltas, format_number)
 from perf import _cpu_utils as cpu_utils
 from perf import _utils as utils
-from perf import tests
 from perf.tests import mock
 from perf.tests import unittest
 
@@ -235,73 +232,6 @@ class CPUToolsTests(unittest.TestCase):
         format_cpus_as_mask = cpu_utils.format_cpus_as_mask
         self.assertEqual(format_cpus_as_mask({4, 5, 6, 7}),
                          '000000f0')
-
-
-class MiscTests(unittest.TestCase):
-    def test_format_metadata(self):
-        self.assertEqual(perf.format_metadata('loops', 2 ** 24),
-                         '2^24')
-
-    def test_python_implementation(self):
-        name = perf.python_implementation()
-        self.assertIsInstance(name, str)
-        self.assertRegex(name, '^[a-z]+$')
-
-    def test_python_has_jit(self):
-        jit = perf.python_has_jit()
-        self.assertIsInstance(jit, bool)
-
-    @unittest.skipUnless(hasattr(os, 'symlink'), 'need os.symlink')
-    def test_abs_executable(self):
-        with tests.temporary_file() as tmpname:
-            tmpname = os.path.realpath(tmpname)
-
-            try:
-                os.symlink(sys.executable, tmpname)
-            except (OSError, NotImplementedError):
-                self.skipTest("os.symlink() failed")
-
-            self.assertEqual(utils.abs_executable(tmpname),
-                             tmpname)
-
-    def test_parse_run_list(self):
-        parse_run_list = utils.parse_run_list
-
-        with self.assertRaises(ValueError):
-            parse_run_list('')
-        with self.assertRaises(ValueError):
-            parse_run_list('0')
-        self.assertEqual(parse_run_list('1'),
-                         [0])
-        self.assertEqual(parse_run_list('1-2,5-6'),
-                         [0, 1, 4, 5])
-        self.assertEqual(parse_run_list('1,3,7'),
-                         [0, 2, 6])
-
-        # tolerate spaces
-        self.assertEqual(parse_run_list(' 1 , 2 '),
-                         [0, 1])
-
-        # errors
-        self.assertRaises(ValueError, parse_run_list, 'x')
-        self.assertRaises(ValueError, parse_run_list, '1,')
-
-    def test_setup_version(self):
-        import setup
-        self.assertEqual(perf.__version__, setup.VERSION)
-
-    def test_doc_version(self):
-        doc_path = os.path.join(os.path.dirname(__file__), '..', '..', 'doc')
-        doc_path = os.path.realpath(doc_path)
-
-        old_path = sys.path[:]
-        try:
-            sys.path.insert(0, doc_path)
-            import conf
-            self.assertEqual(perf.__version__, conf.version)
-            self.assertEqual(perf.__version__, conf.release)
-        finally:
-            sys.path[:] = old_path
 
 
 if __name__ == "__main__":